Bugs fixed

+++ logic/src/main/java/ua/com/tracksee/logic/address/AddressEntityBean.java
@@ -30,7 +30,7 @@ public class AddressEntityBean {
      * @param name -
      * @return -
      */
-    public AddressEntity getAddressByUserId(long userId, String name) {
+    public AddressEntity getAddressByUserId(Integer userId, String name) {
         return addressDAO.getAddressByUserId(userId, name);
     }
 
@@ -39,7 +39,7 @@ public class AddressEntityBean {
      * @param userId -
      * @return p-
      */
-    public List getAllAddressesByUserId(long userId) {
+    public List getAllAddressesByUserId(Integer userId) {
         return addressDAO.getAllAddressesByUserId(userId);
     }
