+++ storage/src/main/java/ua/com/tracksee/dao/AddressDAO.java
@@ -16,8 +16,8 @@ public interface AddressDAO {
     List getAllAddressesByUserId(AddressEntity address);
     List<AddressEntity> getAddresses(int partNumber);
 
-    AddressEntity getAddressByUserId(long userId, String name);
-    List getAllAddressesByUserId(long userId);
+    AddressEntity getAddressByUserId(Integer userId, String name);
+    List getAllAddressesByUserId(Integer userId);
 
 }
